pyproject.toml
[tool.poetry]
name = "hackathonopenai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
geopandas = "^1.0.1"
wget = "^3.2"
matplotlib = "^3.9.2"
openai = "^1.42.0"
python-dotenv = "^1.0.1"
streamlit = "^1.37.1"
streamlit-folium = "^0.22.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
py_version = 3
profile = "black"

[tool.black]
target-version = ['py312']
